用户登录、菜单查询、加入购物车完成，查看购物车功能待完善

+++ src/main/java/homebar/com/entity/Cart.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 @TableName("carts")
 public class Cart {
-    private Long id;
-    private Long userId;
-    private Long dishId;
+    private Integer id;
+    private Integer userId;
+    private Integer dishId;
 //    private int quantity;
     // getters and setters
 }
